requirements.txt
aioamqp==0.14.0
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.3.0
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-rabbitmq==1.2.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
Django==3.0.8
django-celery-beat==2.0.0
django-common-helpers==0.9.2
django-fsm==2.7.0
django-timezone-field==4.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
isort==4.3.21
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
pamqp==2.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.5.3
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
